requirements.txt
aniso8601==6.0.0
attrs==19.1.0
click==7.0
flask-restplus==0.12.1
flask==1.0.3
flask_marshmallow==0.14.0
itsdangerous==1.1.0
jinja2==2.10.1
jsonschema==3.0.1
markupsafe==1.1.1
pyrsistent==0.15.2
pytz==2019.1
six==1.12.0
werkzeug==0.15.5
python-dotenv==0.14.0
sqlalchemy==1.3.19
marshmallow-sqlalchemy==0.24.1
pymysql
beautifulsoup4
requests
